requirements.txt
langchain-mcp-adapters
fastmcp
langchain
langchain-openai
langchain-deepseek
langchain-groq
langchain-neo4j
neo4j>=5.0.0
langgraph
langgraph-checkpoint-redis
fastapi
strawberry-graphql
redis
pdfplumber
gunicorn
uvicorn[standard]
https://github.com/toon-format/toon-python.git
